fix problem when UserPresentReceiver stopped receiving ACTION_USER_PRESENT
Dynamically registered BroadcastReceiver for Intent.ACTION_USER_PRESENT stops receiving intents after some time on Moto G (Android 4.4).
So we will be using statically published receiver in AndroidManifest.xml and enable/disable it as we need.
Plus add button to debug section for dynamically registering this type of receiver to test it on other devices. It will show a Toast message when Intent is received.